refactor(dashboard): share the per-node trends loader between chart widgets
InfraGlobalChartWidget carried its own connection filter, its own trends
loader and its own per-node colour palette. A second chart widget would
have had to copy all three, so they move to their own modules and the
existing widget consumes them.

Behaviour is unchanged and its tests are untouched. The loader gains one
option: the value a missing metric takes. It stays 0 for ratios that are
meaningful at zero, and becomes null for metrics whose absence must stay
distinguishable from a real zero.
